Cognata

+++ train.py
@@ -1,22 +1,56 @@
 """
-@author: Viet Nguyen <nhviet1009@gmail.com>
+Modifications by MLCommons from SSD-Pytorch (https://github.com/uvipen/SSD-pytorch) author: Viet Nguyen (nhviet1009@gmail.com)
+Copyright 2024 MLCommons Association and Contributors
+
+MIT License
+
+Copyright (c) 2021 Viet Nguyen
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
 """
 import os
 import shutil
+import importlib
 from argparse import ArgumentParser
 
 import torch
 from torch.optim.lr_scheduler import MultiStepLR
 from torch.utils.data import DataLoader
-from torch.utils.tensorboard import SummaryWriter
 
 from src.model import SSD, SSDLite, ResNet, MobileNetV2
 from src.utils import generate_dboxes, Encoder, coco_classes
 from src.transform import SSDTransformer
 from src.loss import Loss
-from src.process import train, evaluate
-from src.dataset import collate_fn, CocoDataset
+from src.process import train, evaluate, cognata_eval
+from src.dataset import collate_fn, CocoDataset, Cognata, prepare_cognata, train_val_split
+from torchinfo import summary 
 
+def cleanup():
+    torch.distributed.destroy_process_group()
+
+def prepare(dataset, params, rank, world_size, val=False):
+    sampler = torch.utils.data.distributed.DistributedSampler(dataset, num_replicas=world_size, rank=rank, shuffle=False, drop_last=False)
+    if val:
+        sampler = None
+    dataloader = DataLoader(dataset, sampler=sampler, **params)
+    
+    return dataloader
 
 def get_args():
     parser = ArgumentParser(description="Implementation of SSD")
@@ -38,50 +72,62 @@ def get_args():
     parser.add_argument("--momentum", type=float, default=0.9, help="momentum argument for SGD optimizer")
     parser.add_argument("--weight-decay", type=float, default=0.0005, help="momentum argument for SGD optimizer")
     parser.add_argument("--nms-threshold", type=float, default=0.5)
-    parser.add_argument("--num-workers", type=int, default=4)
-
-    parser.add_argument('--local_rank', default=0, type=int,
+    parser.add_argument("--num-workers", type=int, default=0)
+    parser.add_argument('--local-rank', default=0, type=int,
                         help='Used for multi-process training. Can either be manually set ' +
                              'or automatically set by using \'python -m multiproc\'.')
+    parser.add_argument("--dataset", default='Cognata', type=str)
+    parser.add_argument("--config", default='config', type=str)
+    parser.add_argument("--save-name", default='SSD', type=str)
+    parser.add_argument("--pretrained-model", type=str, default="trained_models/SSD.pth")
+    parser.add_argument("--checkpoint-freq", type=int, default=1)
+    
     args = parser.parse_args()
     return args
 
 
-def main(opt):
-    if torch.cuda.is_available():
-        torch.distributed.init_process_group(backend='nccl', init_method='env://')
-        num_gpus = torch.distributed.get_world_size()
-        torch.cuda.manual_seed(123)
-    else:
-        torch.manual_seed(123)
-        num_gpus = 1
-
-    train_params = {"batch_size": opt.batch_size * num_gpus,
-                    "shuffle": True,
-                    "drop_last": False,
+def main(rank, opt, world_size):
+    train_params = {"batch_size": opt.batch_size,
                     "num_workers": opt.num_workers,
                     "collate_fn": collate_fn}
 
-    test_params = {"batch_size": opt.batch_size * num_gpus,
-                   "shuffle": False,
-                   "drop_last": False,
+    test_params = {"batch_size": opt.batch_size,
                    "num_workers": opt.num_workers,
                    "collate_fn": collate_fn}
 
+    config = importlib.import_module('config.' + opt.config)
+    image_size = config.model['image_size']
+    num_classes = len(coco_classes)
     if opt.model == "ssd":
-        dboxes = generate_dboxes(model="ssd")
-        model = SSD(backbone=ResNet(), num_classes=len(coco_classes))
+        dboxes = generate_dboxes(config.model, model="ssd")
     else:
         dboxes = generate_dboxes(model="ssdlite")
+    if opt.dataset == 'Cognata':
+        folders = config.dataset['folders']
+        cameras = config.dataset['cameras']
+        ignore_classes = [2, 25, 31]
+        if 'ignore_classes' in config.dataset:
+            ignore_classes = config.dataset['ignore_classes']
+        files, label_map, label_info = prepare_cognata(opt.data_path, folders, cameras, ignore_classes)
+        files = train_val_split(files)
+        train_set = Cognata(label_map, label_info, files['train'], ignore_classes, SSDTransformer(dboxes, image_size, val=False))
+        test_set = Cognata(label_map, label_info, files['val'], ignore_classes, SSDTransformer(dboxes, image_size, val=True))
+        num_classes = len(label_map.keys())
+        print(label_map)
+        print(label_info)
+    elif opt.dataset == 'Coco':
+        train_set = CocoDataset(opt.data_path, 2017, "train", SSDTransformer(dboxes, image_size, val=False))
+        test_set = CocoDataset(opt.data_path, 2017, "val", SSDTransformer(dboxes, image_size, val=True))
+    if opt.model == "ssd":
+        model = SSD(config.model, backbone=ResNet(config.model, pretrained=True), num_classes=num_classes)
+    else:
         model = SSDLite(backbone=MobileNetV2(), num_classes=len(coco_classes))
-    train_set = CocoDataset(opt.data_path, 2017, "train", SSDTransformer(dboxes, (300, 300), val=False))
-    train_loader = DataLoader(train_set, **train_params)
-    test_set = CocoDataset(opt.data_path, 2017, "val", SSDTransformer(dboxes, (300, 300), val=True))
-    test_loader = DataLoader(test_set, **test_params)
+    train_loader = prepare(train_set, train_params, rank, world_size)
+    test_loader = prepare(test_set, test_params, rank, world_size, val=True)
 
     encoder = Encoder(dboxes)
 
-    opt.lr = opt.lr * num_gpus * (opt.batch_size / 32)
+    opt.lr = opt.lr * world_size * (opt.batch_size / 32)
     criterion = Loss(dboxes)
 
     optimizer = torch.optim.SGD(model.parameters(), lr=opt.lr, momentum=opt.momentum,
@@ -101,21 +147,23 @@ def main(opt):
             from torch.nn.parallel import DistributedDataParallel as DDP
         # It is recommended to use DistributedDataParallel, instead of DataParallel
         # to do multi-GPU training, even if there is only a single node.
-        model = DDP(model)
+        model = model.to(rank)
+        model = DDP(model, device_ids=[rank], output_device=rank)
 
 
-    if os.path.isdir(opt.log_path):
-        shutil.rmtree(opt.log_path)
-    os.makedirs(opt.log_path)
+    if rank == 0:
+        if os.path.isdir(opt.log_path):
+            shutil.rmtree(opt.log_path)
+        os.makedirs(opt.log_path)
 
     if not os.path.isdir(opt.save_folder):
-        os.makedirs(opt.save_folder)
-    checkpoint_path = os.path.join(opt.save_folder, "SSD.pth")
+        if rank == 0:
+            os.makedirs(opt.save_folder)
+    checkpoint_path = os.path.join(opt.save_folder, opt.save_name)
 
-    writer = SummaryWriter(opt.log_path)
 
-    if os.path.isfile(checkpoint_path):
-        checkpoint = torch.load(checkpoint_path)
+    if os.path.isfile(opt.pretrained_model):
+        checkpoint = torch.load(opt.pretrained_model)
         first_epoch = checkpoint["epoch"] + 1
         model.module.load_state_dict(checkpoint["model_state_dict"])
         scheduler.load_state_dict(checkpoint["scheduler"])
@@ -124,16 +172,21 @@ def main(opt):
         first_epoch = 0
 
     for epoch in range(first_epoch, opt.epochs):
-        train(model, train_loader, epoch, writer, criterion, optimizer, scheduler, opt.amp)
-        evaluate(model, test_loader, epoch, writer, encoder, opt.nms_threshold)
-
+        train(model, train_loader, epoch, criterion, optimizer, scheduler, opt.amp)
         checkpoint = {"epoch": epoch,
                       "model_state_dict": model.module.state_dict(),
                       "optimizer": optimizer.state_dict(),
                       "scheduler": scheduler.state_dict()}
-        torch.save(checkpoint, checkpoint_path)
+        if (rank == 0) and ((epoch+1)%opt.checkpoint_freq == 0):
+            torch.save(checkpoint, checkpoint_path + '_ep' + str(epoch+1) + '.pth')
+        torch.distributed.barrier()
+    cleanup()
 
 
 if __name__ == "__main__":
     opt = get_args()
-    main(opt)
+    torch.distributed.init_process_group("nccl", init_method='env://')
+    world_size = torch.distributed.get_world_size()
+    rank = torch.distributed.get_rank()
+    torch.cuda.set_device(rank)
+    main(rank, opt, world_size)
